lay out new player api
